Cross Section Degree 계산 수정
- Int => double 로 변경

+++ RaywattApp/RaywattApp/ViewModels/Review3dViewModel.cs
@@ -258,7 +258,7 @@ namespace RaywattApp.ViewModels
 
                     double pointXDiff = crossSectionCenter.X - headerSidePointDiff.X;
                     double pointYDiff = crossSectionCenter.Y - headerSidePointDiff.Y;
-                    indicatorDiffDegree = (int)(Math.Atan2(pointYDiff, pointXDiff) * 180 / Math.PI) - Degree;
+                    indicatorDiffDegree = (Math.Atan2(pointYDiff, pointXDiff) * 180 / Math.PI) - Degree;
                     indicator.IsCrossSectionClicked = false;
                 }
 
@@ -274,7 +274,7 @@ namespace RaywattApp.ViewModels
 
                 double pointX = crossSectionCenter.X - headerSidePoint.X;
                 double pointY = crossSectionCenter.Y - headerSidePoint.Y;
-                Degree = (int)(Math.Atan2(pointY, pointX) * 180 / Math.PI) - indicatorDiffDegree;
+                Degree = (Math.Atan2(pointY, pointX) * 180 / Math.PI) - indicatorDiffDegree;
             }
         }
