+++ RaywattApp/RaywattApp/ViewModels/ReviewViewModel.cs
@@ -330,7 +330,7 @@ namespace RaywattApp.ViewModels
 
                     double pointXDiff = crossSectionCenter.X - headerSidePointDiff.X;
                     double pointYDiff = crossSectionCenter.Y - headerSidePointDiff.Y;
-                    indicatorDiffDegree = (int)(Math.Atan2(pointYDiff, pointXDiff) * 180 / Math.PI) - Degree;
+                    indicatorDiffDegree = (Math.Atan2(pointYDiff, pointXDiff) * 180 / Math.PI) - Degree;
                     indicator.IsCrossSectionClicked = false;
                 }
 
@@ -346,7 +346,7 @@ namespace RaywattApp.ViewModels
 
                 double pointX = crossSectionCenter.X - headerSidePoint.X;
                 double pointY = crossSectionCenter.Y - headerSidePoint.Y;
-                Degree = (int)(Math.Atan2(pointY, pointX) * 180 / Math.PI) - indicatorDiffDegree;
+                Degree = (Math.Atan2(pointY, pointX) * 180 / Math.PI) - indicatorDiffDegree;
             }
         }
